Large refactor of scanners
