+++ src/main/java/me/cortex/jarscanner/Main.java
@@ -5,7 +5,7 @@ import picocli.CommandLine;
 /**
  * Main entry point for the CLI
  *
- * @see RootScannerTask CLI command outline.
+ * @see RootScanCommand CLI command outline.
  */
 public final class Main {
     private Main() {
@@ -13,7 +13,7 @@ public final class Main {
 
     public static void main(String[] args) {
         // Delegate app args to picocli, which will populate the root-scanner parameters and options
-        int exitCode = new CommandLine(new RootScannerTask()).execute(args);
+        int exitCode = new CommandLine(new RootScanCommand()).execute(args);
         System.exit(exitCode);
     }
 }
