+++ src/main/java/me/cortex/jarscanner/RootScanCommand.java
@@ -2,6 +2,9 @@ package me.cortex.jarscanner;
 
 import me.cortex.jarscanner.detection.Detection;
 import me.cortex.jarscanner.detection.DetectionManager;
+import me.cortex.jarscanner.scanner.RootScanner;
+import me.cortex.jarscanner.scanner.Scanner;
+import me.cortex.jarscanner.scanner.summary.RootScanSummary;
 import org.slf4j.Logger;
 import picocli.CommandLine;
 
@@ -20,8 +23,8 @@ import static org.slf4j.LoggerFactory.getLogger;
  */
 @CommandLine.Command(/*name = "scan",*/ mixinStandardHelpOptions = true, version = "1.0.0",
         description = "Scans the specified directory for malicious signatures.")
-public class RootScannerTask implements Callable<RootScanSummary> {
-    private static final Logger logger = getLogger(RootScannerTask.class);
+public class RootScanCommand implements Callable<RootScanSummary> {
+    private static final Logger logger = getLogger(RootScanCommand.class);
 
     @CommandLine.Parameters(index = "0", description = "The directory to scan.")
     private File file;
@@ -40,6 +43,8 @@ public class RootScannerTask implements Callable<RootScanSummary> {
         logger.info("Running detectors: {}", detectionsToScanFor.stream().map(Detection::getName).collect(Collectors.joining(", ")));
 
         // Run new scanner
-        return new RootScanner(this.file.toPath(), this.threads, this.timeoutMinutes, detectionsToScanFor).run();
+        Scanner<RootScanSummary> scanner = new RootScanner(this.threads, this.timeoutMinutes, detectionsToScanFor);
+
+        return scanner.runScan(this.file.toPath());
     }
 }
