--- src/main/java/me/cortex/jarscanner/RootScanSummary.java
@@ -1,12 +1,0 @@
-package me.cortex.jarscanner;
-
-import java.util.List;
-
-public class RootScanSummary {
-	// TODO: outline
-
-	public static RootScanSummary fromJarScans(List<JarScanSummary> jarScans) {
-		// TODO: Summarize data
-		return null;
-	}
-}
