+++ src/main/java/me/cortex/jarscanner/scanner/summary/RootScanSummary.java
@@ -1,0 +1,12 @@
+package me.cortex.jarscanner.scanner.summary;
+
+import java.util.List;
+
+public class RootScanSummary implements ScanSummary {
+    // TODO: outline
+
+    public static RootScanSummary fromJarScans(List<JarScanSummary> jarScans) {
+        // TODO: Summarize data
+        return null;
+    }
+}
