+++ src/main/java/me/cortex/jarscanner/scanner/summary/ScanSummary.java
@@ -1,0 +1,5 @@
+package me.cortex.jarscanner.scanner.summary;
+
+public interface ScanSummary {
+    // TODO: 2023-06-08 Add shit here
+}
